pyproject.toml
[tool.poetry]
name = "lammps_executor"
version = "0.1.0"
description = ""
authors = ["feugmo <giresse.feugmo@gmail.com>"]
readme ="README.rst"


[tool.poetry.dependencies]
python = ">=3.8,<3.11"
aiida-lammps = {git = "https://github.com/Feugmo/aiida-lammps"}
clusterX = "^1.0.0-alpha.8"
pandas = "^1.4.3"
pymatgen = "^2022.5.26"
qmpy-rester = "^0.2.0"
mendeleev = "^0.9.0"
psycopg2-binary = "2.8.6"
loguru = "^0.6.0"
mp-api = "^0.24.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
poethepoet = "^0.13.1"
pre-commit = "^2.19.0"
isort = "^5.10.1"
black = "22.3"
toc = "^0.0.11"
flake8 = "^4.0.1"
autopep8 = "^1.6.0"

[tool.poetry.extras]
test = ["pytest"]
docs = ["sphinx", "sphinx_rtd_theme"]



[tool.poe.tasks]
lammps = "conda install -c conda-forge lammps"

# run

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
